@@ -18,51 +18,52 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 0.3.9+28
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  web_socket_channel: ^1.2.0
-  crypto: ^2.0.3
-  bip39: ^1.0.3
+  web_socket_channel: ^2.1.0
+  crypto: ^3.0.1
+  bip39: ^1.0.6
   bitcoin_flutter:
     git:
       url: https://github.com/peercoin/bitcoin_flutter.git
-  hive: ^1.4.4+1
-  hive_flutter: ^0.3.1
-  hive_generator: ^0.8.2
-  flutter_secure_storage: ^4.1.0
-  shared_preferences: ^2.0.5
+      ref: null_safety
+  hive: ^2.0.4
+  hive_flutter: ^1.0.0
+  hive_generator: ^1.1.0
+  flutter_secure_storage: ^4.2.0
+  shared_preferences: ^2.0.6
   provider: ^5.0.0
-  share: ^0.6.5+4
-  qr_flutter: ^3.2.0
+  share: ^2.0.4
+  qr_flutter: ^4.0.0
   intl: ^0.17.0
-  step_progress_indicator: ^0.2.5+8
+  step_progress_indicator: ^1.0.0
   qr_code_scanner: ^0.4.0
-  url_launcher: ^6.0.2
-  flutter_local_notifications: ^5.0.0
+  url_launcher: ^6.0.4
+  flutter_local_notifications: ^6.0.0
   flutter_screen_lock: ^4.0.4+1
-  local_auth: ^1.1.5
-  package_info: ^2.0.0
+  local_auth: ^1.1.6
+  package_info: ^2.0.2
   mailto: ^2.0.0
-  flutter_search_bar: ^2.1.0
+  flutter_search_bar: ^3.0.0-dev.1
   flutter_slidable: ^0.6.0
-  flutter_typeahead: ^3.1.2
-  connectivity: ^3.0.3
+  flutter_typeahead: ^3.1.3
+  connectivity: ^3.0.6
   pedantic: ^1.11.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.3
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  build_runner: ^1.11.1
-  flutter_launcher_icons: "^0.8.0"
-  flutter_native_splash: ^0.3.2
+  # test: ^1.17.5
+  build_runner: ^2.0.4
+  flutter_launcher_icons: ^0.9.0
+  flutter_native_splash: ^1.1.8+4
   
 flutter_icons:
   android: "launcher_icon"
