@@ -16,17 +16,17 @@ class Coin {
   final String genesisHash;
 
   Coin({
-    @required this.name,
-    @required this.displayName,
-    @required this.letterCode,
-    @required this.iconPath,
-    @required this.iconPathTransparent,
-    @required this.uriCode,
-    @required this.networkType,
-    @required this.fractions,
-    @required this.minimumTxValue,
-    @required this.feePerKb,
-    @required this.explorerTxDetailUrl,
-    @required this.genesisHash,
+    required this.name,
+    required this.displayName,
+    required this.letterCode,
+    required this.iconPath,
+    required this.iconPathTransparent,
+    required this.uriCode,
+    required this.networkType,
+    required this.fractions,
+    required this.minimumTxValue,
+    required this.feePerKb,
+    required this.explorerTxDetailUrl,
+    required this.genesisHash,
   });
 }
