@@ -5,44 +5,44 @@ part 'server.g.dart';
 @HiveType(typeId: 6)
 class Server extends HiveObject {
   @HiveField(0)
-  String _label;
+  String? _label;
 
   @HiveField(1)
-  String address;
+  String? address;
 
   @HiveField(2)
-  bool connectable = true;
+  bool? connectable = true;
 
   @HiveField(3)
-  bool userGenerated;
+  bool? userGenerated;
 
   @HiveField(4)
-  String _donationAddress;
+  String? _donationAddress;
 
   @HiveField(5)
-  String _serverBanner;
+  String? _serverBanner;
 
   @HiveField(6)
-  DateTime _lastConnection;
+  DateTime? _lastConnection;
 
   @HiveField(7)
   // ignore: prefer_final_fields
-  bool _hidden = false;
+  bool? _hidden = false;
 
   @HiveField(8)
-  int priority;
+  int? priority;
 
   Server({
-    @required this.address,
-    @required this.priority,
-    @required this.userGenerated,
+    required this.address,
+    required this.priority,
+    required this.userGenerated,
   });
 
-  String get getAddress {
+  String? get getAddress {
     return address;
   }
 
-  bool get hidden {
+  bool? get hidden {
     return _hidden;
   }
 
