@@ -5,41 +5,41 @@ part 'wallettransaction.g.dart';
 @HiveType(typeId: 3)
 class WalletTransaction extends HiveObject {
   @HiveField(0)
-  final String txid;
+  final String? txid;
   @HiveField(1)
-  int timestamp;
+  int? timestamp;
   @HiveField(2)
-  final int value;
+  final int? value;
   @HiveField(3)
-  final int fee;
+  final int? fee;
   @HiveField(4)
-  final String address;
+  final String? address;
   @HiveField(5)
-  final String direction;
+  final String? direction;
   @HiveField(6)
-  int confirmations = 0;
+  int? confirmations = 0;
   @HiveField(7)
-  bool broadCasted = true;
+  bool? broadCasted = true;
   @HiveField(8)
-  String broadcastHex = '';
+  String? broadcastHex = '';
 
   WalletTransaction({
-    @required this.txid,
-    @required this.timestamp,
-    @required this.value,
-    @required this.fee,
-    @required this.address,
-    @required this.direction,
-    @required this.broadCasted,
-    @required this.broadcastHex,
-    @required this.confirmations,
+    required this.txid,
+    required this.timestamp,
+    required this.value,
+    required this.fee,
+    required this.address,
+    required this.direction,
+    required this.broadCasted,
+    required this.broadcastHex,
+    required this.confirmations,
   });
 
-  set newTimestamp(int newTime) {
+  set newTimestamp(int? newTime) {
     timestamp = newTime;
   }
 
-  set newConfirmations(int newConfirmations) {
+  set newConfirmations(int? newConfirmations) {
     confirmations = newConfirmations;
   }
 
