@@ -17,16 +17,16 @@ class ServerAdapter extends TypeAdapter<Server> {
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
     return Server(
-      address: fields[1] as String,
-      priority: fields[8] as int,
-      userGenerated: fields[3] as bool,
+      address: fields[1] as String?,
+      priority: fields[8] as int?,
+      userGenerated: fields[3] as bool?,
     )
-      .._label = fields[0] as String
-      ..connectable = fields[2] as bool
-      .._donationAddress = fields[4] as String
-      .._serverBanner = fields[5] as String
-      .._lastConnection = fields[6] as DateTime
-      .._hidden = fields[7] as bool;
+      .._label = fields[0] as String?
+      ..connectable = fields[2] as bool?
+      .._donationAddress = fields[4] as String?
+      .._serverBanner = fields[5] as String?
+      .._lastConnection = fields[6] as DateTime?
+      .._hidden = fields[7] as bool?;
   }
 
   @override
