@@ -19,13 +19,13 @@ Map<String, Coin> availableCoins = AvailableCoins().availableCoins;
 List activeCoins = [];
 
 class _NewWalletScreenState extends State<NewWalletScreen> {
-  String _coin = '';
+  String? _coin = '';
   bool _initial = true;
 
   Future<void> addWallet(ctx) async {
     try {
       await Provider.of<ActiveWallets>(context, listen: false).addWallet(_coin,
-          availableCoins[_coin].displayName, availableCoins[_coin].letterCode);
+          availableCoins[_coin!]!.displayName, availableCoins[_coin!]!.letterCode);
       var prefs =
           await Provider.of<UnencryptedOptions>(context, listen: false).prefs;
 
@@ -40,8 +40,8 @@ class _NewWalletScreenState extends State<NewWalletScreen> {
       ScaffoldMessenger.of(context).showSnackBar(SnackBar(
         content: Text(
           _coin == ''
-              ? AppLocalizations.instance.translate('select_coin')
-              : AppLocalizations.instance.translate('add_coin_failed'),
+              ? AppLocalizations.instance.translate('select_coin')!
+              : AppLocalizations.instance.translate('add_coin_failed')!,
           textAlign: TextAlign.center,
         ),
         duration: Duration(seconds: 2),
@@ -53,8 +53,8 @@ class _NewWalletScreenState extends State<NewWalletScreen> {
   void didChangeDependencies() async {
     if (_initial) {
       var _appSettings = Provider.of<AppSettings>(context, listen: false);
-      if (_appSettings.authenticationOptions['newWallet']) {
-        await Auth.requireAuth(context, _appSettings.biometricsAllowed);
+      if (_appSettings.authenticationOptions!['newWallet']!) {
+        await Auth.requireAuth(context, _appSettings.biometricsAllowed!);
       }
       setState(() {
         _initial = false;
@@ -82,7 +82,7 @@ class _NewWalletScreenState extends State<NewWalletScreen> {
 
     return Scaffold(
       appBar: AppBar(
-        title: Text(AppLocalizations.instance.translate('add_new_wallet')),
+        title: Text(AppLocalizations.instance.translate('add_new_wallet')!),
         actions: [
           if (actualAvailableWallets.isNotEmpty)
             Padding(
@@ -97,7 +97,7 @@ class _NewWalletScreenState extends State<NewWalletScreen> {
         child: actualAvailableWallets.isEmpty
             ? Center(
                 child:
-                    Text(AppLocalizations.instance.translate('no_new_wallet')),
+                    Text(AppLocalizations.instance.translate('no_new_wallet')!),
               )
             : Column(children: [
                 Expanded(
@@ -119,18 +119,18 @@ class _NewWalletScreenState extends State<NewWalletScreen> {
                               child: Image.asset(
                                   AvailableCoins()
                                       .getSpecificCoin(availableCoins[
-                                              actualAvailableWallets[item]]
-                                          .name)
+                                              actualAvailableWallets[item]]!
+                                          .name)!
                                       .iconPath,
                                   width: 20),
                             ),
                             title: Text(
-                                availableCoins[actualAvailableWallets[item]]
+                                availableCoins[actualAvailableWallets[item]]!
                                     .displayName),
                             leading: Radio(
                               value: actualAvailableWallets[item],
                               groupValue: _coin,
-                              onChanged: (value) {
+                              onChanged: (dynamic value) {
                                 setState(
                                   () {
                                     _coin = value;
