@@ -17,16 +17,16 @@ class TransactionDetails extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final args = ModalRoute.of(context).settings.arguments as List;
+    final args = ModalRoute.of(context)!.settings.arguments as List;
     final WalletTransaction _tx = args[0];
     final CoinWallet _coinWallet = args[1];
     final baseUrl =
-        AvailableCoins().getSpecificCoin(_coinWallet.name).explorerTxDetailUrl;
+        AvailableCoins().getSpecificCoin(_coinWallet.name)!.explorerTxDetailUrl;
 
     return Scaffold(
       appBar: AppBar(
           title: Text(
-        AppLocalizations.instance.translate('transaction_details'),
+        AppLocalizations.instance.translate('transaction_details')!,
       )),
       body: ListView(
         padding: EdgeInsets.all(20),
@@ -34,21 +34,21 @@ class TransactionDetails extends StatelessWidget {
           Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              Text(AppLocalizations.instance.translate('id'),
+              Text(AppLocalizations.instance.translate('id')!,
                   style: TextStyle(fontWeight: FontWeight.bold)),
-              SelectableText(_tx.txid)
+              SelectableText(_tx.txid!)
             ],
           ),
           Divider(),
           Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              Text(AppLocalizations.instance.translate('time'),
+              Text(AppLocalizations.instance.translate('time')!,
                   style: TextStyle(fontWeight: FontWeight.bold)),
               SelectableText(_tx.timestamp != null
                   ? DateFormat().format(
-                      DateTime.fromMillisecondsSinceEpoch(_tx.timestamp * 1000))
-                  : AppLocalizations.instance.translate('unconfirmed'))
+                      DateTime.fromMillisecondsSinceEpoch(_tx.timestamp! * 1000))
+                  : AppLocalizations.instance.translate('unconfirmed')!)
             ],
           ),
           Divider(),
@@ -56,31 +56,31 @@ class TransactionDetails extends StatelessWidget {
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               Text(
-                AppLocalizations.instance.translate('tx_value'),
+                AppLocalizations.instance.translate('tx_value')!,
                 style: TextStyle(fontWeight: FontWeight.bold),
               ),
-              SelectableText((_tx.value / 1000000).toString() +
+              SelectableText((_tx.value! / 1000000).toString() +
                   ' ' +
-                  _coinWallet.letterCode)
+                  _coinWallet.letterCode!)
             ],
           ),
           Divider(),
           Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              Text(AppLocalizations.instance.translate('tx_fee'),
+              Text(AppLocalizations.instance.translate('tx_fee')!,
                   style: TextStyle(fontWeight: FontWeight.bold)),
               SelectableText(
-                  (_tx.fee / 1000000).toString() + ' ' + _coinWallet.letterCode)
+                  (_tx.fee! / 1000000).toString() + ' ' + _coinWallet.letterCode!)
             ],
           ),
           Divider(),
           Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              Text(AppLocalizations.instance.translate('tx_address'),
+              Text(AppLocalizations.instance.translate('tx_address')!,
                   style: TextStyle(fontWeight: FontWeight.bold)),
-              SelectableText(_tx.address),
+              SelectableText(_tx.address!),
               // Text("") TODO might add address label here in the future
             ],
           ),
@@ -88,16 +88,16 @@ class TransactionDetails extends StatelessWidget {
           Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              Text(AppLocalizations.instance.translate('tx_direction'),
+              Text(AppLocalizations.instance.translate('tx_direction')!,
                   style: TextStyle(fontWeight: FontWeight.bold)),
-              SelectableText(_tx.direction)
+              SelectableText(_tx.direction!)
             ],
           ),
           Divider(),
           Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              Text(AppLocalizations.instance.translate('tx_confirmations'),
+              Text(AppLocalizations.instance.translate('tx_confirmations')!,
                   style: TextStyle(fontWeight: FontWeight.bold)),
               SelectableText(_tx.confirmations.toString())
             ],
@@ -110,7 +110,7 @@ class TransactionDetails extends StatelessWidget {
                   Icons.search,
                 ),
                 label: Text(
-                  AppLocalizations.instance.translate('tx_view_in_explorer'),
+                  AppLocalizations.instance.translate('tx_view_in_explorer')!,
                 )),
           )
         ],
