@@ -19,7 +19,7 @@ class AppDrawer extends StatelessWidget {
       child: (Column(
         children: [
           AppBar(
-            title: Text(AppLocalizations.instance.translate('app_navigation')),
+            title: Text(AppLocalizations.instance.translate('app_navigation')!),
             automaticallyImplyLeading: false,
           ),
           ListTile(
@@ -28,38 +28,38 @@ class AppDrawer extends StatelessWidget {
             leading: Icon(
               Icons.account_balance_wallet,
             ),
-            title: Text(AppLocalizations.instance.translate('app_wallets')),
+            title: Text(AppLocalizations.instance.translate('app_wallets')!),
           ),
           Divider(),
           ListTile(
               leading: Icon(Icons.app_settings_alt),
-              title: Text(AppLocalizations.instance.translate('app_settings')),
+              title: Text(AppLocalizations.instance.translate('app_settings')!),
               onTap: () => Navigator.of(context)
                   .pushReplacementNamed(Routes.AppSettings)),
           Divider(),
           ListTile(
             leading: Icon(Icons.shopping_basket),
-            title: Text(AppLocalizations.instance.translate('buy_peercoin')),
+            title: Text(AppLocalizations.instance.translate('buy_peercoin')!),
             onTap: () => showDialog(
               context: context,
               builder: (_) => AlertDialog(
                 title: Text(
                   AppLocalizations.instance
-                      .translate('buy_peercoin_dialog_title'),
+                      .translate('buy_peercoin_dialog_title')!,
                 ),
                 content: Text(AppLocalizations.instance
-                    .translate('buy_peercoin_dialog_content')),
+                    .translate('buy_peercoin_dialog_content')!),
                 actions: <Widget>[
                   TextButton.icon(
                       label: Text(AppLocalizations.instance
-                          .translate('server_settings_alert_cancel')),
+                          .translate('server_settings_alert_cancel')!),
                       icon: Icon(Icons.cancel),
                       onPressed: () {
                         Navigator.of(context).pop();
                       }),
                   TextButton.icon(
                     label: Text(AppLocalizations.instance
-                        .translate('jail_dialog_button')),
+                        .translate('jail_dialog_button')!),
                     icon: Icon(Icons.check),
                     onPressed: () {
                       _launchURL('https://ppc.lol/buy');
@@ -74,13 +74,13 @@ class AppDrawer extends StatelessWidget {
 
           ListTile(
               leading: Icon(Icons.info),
-              title: Text(AppLocalizations.instance.translate('about')),
+              title: Text(AppLocalizations.instance.translate('about')!),
               onTap: () =>
                   Navigator.of(context).pushReplacementNamed(Routes.About)),
           Divider(),
           ListTile(
             leading: Icon(Icons.share),
-            title: Text(AppLocalizations.instance.translate('share_app')),
+            title: Text(AppLocalizations.instance.translate('share_app')!),
             onTap: () => Share.share(
                 'https://play.google.com/store/apps/details?id=com.coinerella.peercoin'),
           ) //TODO point to peercoin.net when it's on there
